Add interactive analysis capabilities
This code attempts to break up the sql code in
inst/sql/sql_server/CohortMethod.sql into more manageable pieces which
make an interactive form of analysis easier. The functions are
structured in a rudimentary object-oriented fashion. See
examples/InteractiveAnalysisExample.R for more details.

Currently the sql code is written targeted at the redshift driver
instead of using functions like loadRenderSql to more generally target
all supported database management systems. Furthermore, the sql files
are based off of an earlier version of CohortMethod.sql so the
functionality here is not up to date with the mainline as of this
commit.