Fix schema search path to find vocabulary
The concept_ancestor table is in the vocabulary schema. Apparently the
table is in the mslr_cdm4 schema (or possibly a view), but that is not
the expected behavior and is currently not true for ccae_cdm4. This
commit changes the sql code to refer to the vocabulary schema instead
and works with both the mslr_cdm4 and ccae_cdm4 schemas.