Temporary fix to to vocabulary schema references
In CDM4, the vocabulary schema is independent of the mslr, ccae, etc.
schemas. In CDM5, this will no longer be the case. This commit changes
the implicit references to the vocabulary to explicit ones so as to work
with CDM4. This will no longer be necessary (and hence this commit will
be reverted), once the switch to CDM5 is complete.