Add getCovariateSize method.
This method returns the number of distinct persons and number of
distinct created by the buildCovariates. It is similar to the
getCohortSize method. Currently both of these size methods use a
hard-coded sql query targetted at a redshift database. These need to be
generalized to work with loadRenderTranslateSql at sometime in the
future.